migrate action handler into shortcuts module
