+++ src/input.rs
@@ -1,10 +1,8 @@
-//! Root of the host input pipeline: keyboard, mouse, focus, gesture and binding
-//! primitives, IME, hyperlink-hover, shortcuts, option-as-alt, and the
+//! Root of the host input pipeline: keyboard, mouse, focus, gesture and binding primitives, IME, hyperlink-hover, shortcuts, option-as-alt, and the
 //! per-mode (`default_mode`, `tmux`) dispatchers.
 
 mod bindings;
 pub(crate) mod default_mode;
-mod dispatch;
 pub(crate) mod focus;
 mod gesture;
 pub(crate) mod hyperlink;
@@ -18,9 +16,9 @@ pub(crate) mod tmux;
 
 use crate::{
     input::{
-        default_mode::DefaultHostInputPlugin, dispatch::DispatchPlugin,
-        keyboard::KeyboardInputPlugin, mouse::MouseInputPlugin, option_as_alt::OptionAsAltPlugin,
-        shortcuts::ShortcutsPlugin, tmux::TmuxInputPlugin,
+        default_mode::DefaultHostInputPlugin, keyboard::KeyboardInputPlugin,
+        mouse::MouseInputPlugin, option_as_alt::OptionAsAltPlugin, shortcuts::ShortcutsPlugin,
+        tmux::TmuxInputPlugin,
     },
     system_set::OzmuxSystems,
 };
@@ -51,7 +49,6 @@ impl Plugin for OzmuxInputPlugin {
             OptionAsAltPlugin,
             KeyboardInputPlugin,
             MouseInputPlugin,
-            DispatchPlugin,
             TmuxInputPlugin,
             DefaultHostInputPlugin,
         ))
